requirements.txt
boto3==1.15.17
celery==4.4.7
Django==3.1.2
django-bootstrap4==2.2.0
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-simple-captcha==0.5.12
django-storages==1.10.1
djangorestframework==3.12.1
easy-thumbnails==2.7
gunicorn==20.0.4
ipython==7.18.1
Pillow==7.2.0
pre-commit==2.8.2
psycopg2-binary==2.8.6
pydot==1.4.1
pylint==2.6.0
pyparsing==3.0.0a2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-factoryboy==2.0.3
redis==3.5.3
social-auth-app-django==4.0.0
